requirements.txt
anyio==4.11.0
asttokens==3.0.0
certifi==2025.10.5
click==8.3.0
clip==0.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
icecream==2.1.8
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
PyYAML==6.0.3
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
triton==3.5.0
typer-slim==0.20.0
typing_extensions==4.15.0
